requirements.txt
python-dotenv==1.0.0
requests==2.32.3
pypdf==5.1.0
python-docx==1.1.2
flask==3.0.3
